+++ packages/launchpad/src/lcdapi/distribution.spec.ts
@@ -47,7 +47,7 @@ describe("DistributionExtension", () => {
       const memo = "Test delegation for wasmd";
       const { accountNumber, sequence } = await client.getSequence();
       const signDoc = makeSignDoc([msg], defaultFee, chainId, memo, accountNumber, sequence);
-      const { signed, signature } = await wallet.sign(faucet.address, signDoc);
+      const { signed, signature } = await wallet.signAmino(faucet.address, signDoc);
       const signedTx = makeStdTx(signed, signature);
 
       const result = await client.broadcastTx(signedTx);