+++ packages/proto-signing/types/index.d.ts
@@ -1,6 +1,7 @@
 export { Coin } from "./msgs";
-export { cosmosField } from "./decorator";
-export { Registry } from "./registry";
+export { cosmosField, registered } from "./decorator";
+export { EncodeObject, Registry } from "./registry";
 export { DirectSecp256k1Wallet } from "./directsecp256k1wallet";
 export { decodePubkey, encodePubkey } from "./pubkey";
-export { makeAuthInfo, makeSignBytes } from "./signing";
+export { isOfflineDirectSigner, OfflineDirectSigner, OfflineSigner } from "./signer";
+export { makeAuthInfoBytes, makeSignBytes, makeSignDoc } from "./signing";