+++ packages/stargate/scripts/predefine-proto.sh
@@ -22,11 +22,15 @@ yarn pbjs \
   --force-long \
   "$COSMOS_PROTO_DIR/auth/v1beta1/auth.proto" \
   "$COSMOS_PROTO_DIR/auth/v1beta1/query.proto" \
+  "$COSMOS_PROTO_DIR/bank/v1beta1/bank.proto" \
   "$COSMOS_PROTO_DIR/bank/v1beta1/query.proto" \
+  "$COSMOS_PROTO_DIR/bank/v1beta1/tx.proto" \
   "$COSMOS_PROTO_DIR/base/query/v1beta1/pagination.proto" \
   "$COSMOS_PROTO_DIR/base/v1beta1/coin.proto" \
   "$COSMOS_PROTO_DIR/crypto/multisig/v1beta1/multisig.proto" \
   "$COSMOS_PROTO_DIR/crypto/secp256k1/keys.proto" \
+  "$COSMOS_PROTO_DIR/staking/v1beta1/staking.proto" \
+  "$COSMOS_PROTO_DIR/staking/v1beta1/tx.proto" \
   "$COSMOS_PROTO_DIR/tx/signing/v1beta1/signing.proto" \
   "$COSMOS_PROTO_DIR/tx/v1beta1/tx.proto" \
   "$IBC_PROTO_DIR/core/channel/v1/channel.proto" \
@@ -35,7 +39,12 @@ yarn pbjs \
   "$IBC_PROTO_DIR/core/commitment/v1/commitment.proto" \
   "$IBC_PROTO_DIR/core/connection/v1/connection.proto" \
   "$IBC_PROTO_DIR/core/connection/v1/query.proto" \
-  "$TENDERMINT_PROTO_DIR/crypto/proof.proto"
+  "$TENDERMINT_PROTO_DIR/crypto/keys.proto" \
+  "$TENDERMINT_PROTO_DIR/crypto/proof.proto" \
+  "$TENDERMINT_PROTO_DIR/libs/bits/types.proto" \
+  "$TENDERMINT_PROTO_DIR/types/types.proto" \
+  "$TENDERMINT_PROTO_DIR/types/validator.proto" \
+  "$TENDERMINT_PROTO_DIR/version/types.proto"
 
 # Work around https://github.com/protobufjs/protobuf.js/issues/1477
 # shellcheck disable=SC2016