+++ packages/stargate/types/encoding.d.ts
@@ -1,0 +1,1 @@
+export declare function getMsgType(typeUrl: string): string;